Pull out email bug report functionality into AardvarkMailUI
