+++ Sources/Aardvark/Aardvark.swift
@@ -27,17 +27,6 @@ public typealias EmailAddress = String
 @objc
 public class Aardvark : NSObject {
     
-    /// Sets up a two finger press-and-hold gesture recognizer to trigger email bug reports that will be sent to emailAddress. Returns the created bug reporter for convenience.
-    @objc
-    public static func addDefaultBugReportingGestureWithEmailBugReporter(withRecipient emailAddress: EmailAddress) -> ARKEmailBugReporter {
-        let logStore = ARKLogDistributor.default().defaultLogStore
-        let bugReporter = ARKEmailBugReporter(emailAddress: emailAddress, logStore: logStore)
-        
-        UIApplication.shared.addTwoFingerPressAndHoldGestureRecognizerTrigger(with: bugReporter)
-        
-        return bugReporter;
-    }
-    
     /// Creates and returns a gesture recognizer that when triggered will call [bugReporter composeBugReport].
     @nonobjc
     public static func add<GestureRecognizer: UIGestureRecognizer>(bugReporter: ARKBugReporter, triggeringGestureRecognizerClass: GestureRecognizer.Type) -> GestureRecognizer? {
