+++ Sources/AardvarkMailUI/ARKEmailBugReporter.h
