+++ Sources/AardvarkMailUI/AardvarkMailUI.h
@@ -1,0 +1,21 @@
+//
+//  AardvarkMailUI.h
+//  AardvarkMailUI
+//
+//  Created by Nicholas Entin on 11/6/20.
+//  Copyright © 2020 Square, Inc. All rights reserved.
+//
+
+#import <Foundation/Foundation.h>
+
+//! Project version number for AardvarkMailUI.
+FOUNDATION_EXPORT double AardvarkMailUIVersionNumber;
+
+//! Project version string for AardvarkMailUI.
+FOUNDATION_EXPORT const unsigned char AardvarkMailUIVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <AardvarkMailUI/PublicHeader.h>
+
+
+#import <AardvarkMailUI/ARKEmailBugReporter.h>
+#import <AardvarkMailUI/ARKEmailBugReportConfiguration.h>
